feat: Add new pages for color conversion, image preview, URL encoding/decoding, and regex text extraction
This commit adds new pages for color conversion, image preview, URL encoding/decoding, and regex text extraction. Each page is implemented as a separate component in the `apps/utils/app` directory. This enhancement expands the functionality of the application and provides users with more options for manipulating and analyzing data.

Recent repository commits:
- fix: 每行一个
- refactor: Update UI components and configuration files
- refactor: Update UI components and configuration files
- refactor: Update UI components and configuration files
- feat: Add options to modify comma-separated link string
- Update metadata title and description in layout.tsx
- refactor: update page component and add header component
- chore: add utils app
- chore: update package name to "pdx-tools"
- Update dependencies in package.json files
- fix(deps): update nextjs monorepo to v13.5.5 (#19)
- fix(deps): update dependency eslint-plugin-react to v7.33.2 (#28)